feat: getLogger can specify a custom reporter
