+++ packages/log/src/logReporter.ts
@@ -1,24 +1,11 @@
+import { addLogReporterInternal } from './logReporterInternal'
 import { store } from './store';
 import { LogReporter } from './types';
 import { assertLogModeIsNotProduction } from './utils';
 
 export function addLogReporter(reporter: LogReporter) {
   assertLogModeIsNotProduction('addLogReporter')
-  if (reporter.isConsoleReporter) {
-    const i = store.value.reporters.findIndex(r => r.isConsoleReporter)
-    const r = store.value.reporters[i]
-    store.value.reporters.splice(i, 1)
-    if (r.filter) {
-      if (reporter.filter) {
-        const f2 = reporter.filter
-        reporter.filter = entry => r.filter!(entry) && f2(entry)
-      }
-      else {
-        reporter.filter = r.filter
-      }
-    }
-  }
-  store.value.reporters.push(reporter)
+  addLogReporterInternal(reporter)
 }
 
 export function getLogReporter(id: string) {
